fix: tsconfig issues
