+++ packages/newsletter-subscription/index.html
@@ -11,6 +11,6 @@
 
     <div id="newsletter-form" />
 
-    <script type="module" src="src/index.ts"></script>
+    <script type="module" src="src/main.ts"></script>
   </body>
 </html>
