+++ packages/newsletter-subscription/package.json
@@ -19,7 +19,7 @@
   ],
   "main": "./dist/index.cjs",
   "module": "./dist/index.js",
-  "types": "./dist/main.d.ts",
+  "types": "./dist/index.d.ts",
   "files": [
     "dist/"
   ],
