+++ packages/newsletter-subscription/tsconfig.build.json
@@ -1,14 +1,16 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "baseUrl": ".",
-    "rootDir": ".",
+    "baseUrl": "./src",
+    "rootDir": "./src",
     "typeRoots": ["./node_modules/@types", "./nortic.d.ts"],
     "skipLibCheck": true
   },
+  "include": [
+    "./src/**/*.ts"],
   "exclude": [
     "./node_modules/**",
     "./dist/**",
-    "./src/main.ts"
+    "./src/index.ts"
   ]
 }
